requirements.txt
-e .
arrow==1.2.3
beautifulsoup4==4.11.1
fire==0.5.0
mailersend==0.5.0
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.1
rich==13.2.0
requests==2.28.2
